update file compra-routes.js

+++ prueba de node/routes/compra-routers.js
@@ -11,7 +11,7 @@ router.get('/', (req,res)=>{
 // devolver todos los registros de compras
 router.get('/compras', async (req, res) => {
     const compras = [];
-    sql="SELECT * FROM COMPRA";
+    sql="SELECT compraid, p_nombre, p_apellido, nombre_proveedor, compra.fechacompra, tipo_pago.metodo_pago FROM ADMINISTRADOR.COMPRA INNER JOIN ADMINISTRADOR.PROVEEDOR ON compra.proveedorid=proveedor.proveedorid INNER JOIN ADMINISTRADOR.TIPO_PAGO ON tipo_pago.tipopagoid = compra.tipopagoid INNER JOIN ADMINISTRADOR.EMPLEADO ON empleado.empleadoid = compra.empleadoid order by compraid desc";
 
     let result = await BD.Open(sql,[],false);
     console.log(result.rows);
@@ -21,10 +21,10 @@ router.get('/compras', async (req, res) => {
     result.rows.map(compra=>{
         let userSchema = {
             "COMPRAID": compra[0],
-            "EMPLEADOID": compra[1],
-            "PROVEEDORID": compra[2],
-            "FECHACOMPRA": compra[3],
-            "TIPOPAGOID": compra[4]
+            "EMPLEADOID": `${compra[1]} ${compra[2]}` ,
+            "PROVEEDORID": compra[3],
+            "FECHACOMPRA": compra[4],
+            "TIPOPAGOID": compra[5]
         }
         compras.push(userSchema)
     });
@@ -41,7 +41,7 @@ router.get('/compras/:ID', async (req, res) => {
     try {
 
         // Mapear el resultado y enviarlo como respuesta
-        const sql = "SELECT * FROM COMPRA WHERE COMPRAID = :ID"; 
+        const sql = "SELECT * FROM administrador.COMPRA WHERE COMPRAID = :ID"; 
         const result = await BD.Open(sql, [ID], false);
 
         const compra = {
@@ -64,14 +64,14 @@ router.get('/compras/:ID', async (req, res) => {
 //=================================================================================================
 //insertar un nuevo registro
 router.post('/compras', async (req, res) => {
-    const {COMPRAID, EMPLEADOID, PROVEEDORID, FECHACOMPRA, TIPOPAGOID} = req.body;
+    const {EMPLEADOID, PROVEEDORID, FECHACOMPRA, TIPOPAGOID} = req.body;
 
     console.log(req.body)
     
     try {
-        const sql = "INSERT INTO compra (COMPRAID, EMPLEADOID, PROVEEDORID, FECHACOMPRA, TIPOPAGOID) VALUES (:COMPRAID, :EMPLEADOID, :PROVEEDORID, :FECHACOMPRA, :TIPOPAGOID)";
+        const sql = "INSERT INTO administrador.compra ( EMPLEADOID, PROVEEDORID, FECHACOMPRA, TIPOPAGOID) VALUES (:EMPLEADOID, :PROVEEDORID, :FECHACOMPRA, :TIPOPAGOID)";
         console.log('Consulta SQL:', sql);
-        const bindParams = [COMPRAID, EMPLEADOID, PROVEEDORID, FECHACOMPRA, TIPOPAGOID];
+        const bindParams = [EMPLEADOID, PROVEEDORID, FECHACOMPRA, TIPOPAGOID];
 
         let result = await BD.Open(sql, bindParams, true);
         console.log(result);
@@ -94,7 +94,7 @@ router.put('/compras/:ID', async (req, res) => {
 
     try {
         // Actualizar el registro
-        const updateQuery = "UPDATE COMPRA SET EMPLEADOID=:EMPLEADOID, PROVEEDORID=:PROVEEDORID, FECHACOMPRA=:FECHACOMPRA, TIPOPAGOID=:TIPOPAGOID WHERE COMPRAID = :ID"
+        const updateQuery = "UPDATE administrador.COMPRA SET EMPLEADOID=:EMPLEADOID, PROVEEDORID=:PROVEEDORID, FECHACOMPRA=:FECHACOMPRA, TIPOPAGOID=:TIPOPAGOID WHERE COMPRAID = :ID"
         const bindParams = [EMPLEADOID, PROVEEDORID, FECHACOMPRA, TIPOPAGOID, ID];
 
         let result = await BD.Open(updateQuery, bindParams, true);
@@ -116,7 +116,7 @@ router.delete('/compras/:ID', async (req, res) => {
     try {
 
         // Eliminar el registro
-        const deleteQuery = "DELETE FROM COMPRA WHERE COMPRAID = :ID";
+        const deleteQuery = "DELETE FROM administrador.COMPRA WHERE COMPRAID = :ID";
         const bindParams = [ID];
 
         let result = await BD.Open(deleteQuery, bindParams, true);
@@ -126,6 +126,59 @@ router.delete('/compras/:ID', async (req, res) => {
     } catch (error) {
         console.error(error);
         res.status(500).json({ error: "Error al eliminar el registro de la base de datos" });
+    }
+});
+
+
+//==================================================================================
+router.get('/empleadosParaSelect', async (req, res) => {
+    try {
+        const sql = "SELECT empleadoID, p_nombre, p_apellido FROM administrador.EMPLEADO";
+        const result = await BD.Open(sql, [], false);
+
+        const empleados = result.rows.map(emp => ({
+            empleadoID: emp[0],
+            nombreCompleto: `${emp[1]} ${emp[2]}` // Concatenar primer nombre y primer apellido
+        }));
+
+        res.status(200).json({ empleados });
+    } catch (error) {
+        console.error(error);
+        res.status(500).json({ error: "Error al obtener los datos de empleados" });
+    }
+});
+//=========================================================
+router.get('/proveedoresParaSelect', async (req, res) => {
+    try {
+        const sql = "SELECT * FROM administrador.proveedor";
+        const result = await BD.Open(sql, [], false);
+
+        const proveedores = result.rows.map(prove => ({
+            ID: prove[0],
+            NombreProveedor: prove[1],
+        }));
+
+        res.status(200).json({ proveedores });
+    } catch (error) {
+        console.error(error);
+        res.status(500).json({ error: "Error al obtener los datos de proveedores" });
+    }
+});
+//=============================================
+router.get('/tipoPagoParaSelect', async (req, res) => {
+    try {
+        const sql = "SELECT * FROM administrador.tipo_pago";
+        const result = await BD.Open(sql, [], false);
+
+        const tipoPago = result.rows.map(tPago => ({
+            tipoPagoID: tPago[0],
+            metodoPago: tPago[1]
+        }));
+
+        res.status(200).json({ tipoPago });
+    } catch (error) {
+        console.error(error);
+        res.status(500).json({ error: "Error al obtener los datos de tipo de pago" });
     }
 });
